prov/verbs: Move DGRAM endpoint code to the root folder of verbs provider
